+++ src/platform_impl/orbital/window.rs
@@ -262,6 +262,9 @@ impl Window {
     pub fn set_transparent(&self, _transparent: bool) {}
 
     #[inline]
+    pub fn set_blur(&self, _blur: bool) {}
+
+    #[inline]
     pub fn set_visible(&self, _visibility: bool) {}
 
     #[inline]