+++ CHANGELOG.md
@@ -28,7 +28,9 @@ And please only add new entries to the top of this list, right below the `# Unre
 - **Breaking:** Change default `ControlFlow` from `Poll` to `Wait`.
 - **Breaking:** remove `DeviceEvent::Text`.
 - On Android, fix `DeviceId` to contain device id's.
-- Add `Window::show_window_menu` to request a titlebar/system menu; implemented on Windows for now. 
+- Add `Window::set_blur` to request a blur behind the window; implemented on Wayland for now.
+- On Web, fix `ControlFlow::WaitUntil` to never wake up **before** the given time.
+- Add `Window::show_window_menu` to request a titlebar/system menu; implemented on Windows for now.
 
 # 0.29.1-beta